taskbackend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
openpyxl 
django-cors-headers
python-decouple
et_xmlfile
PyJWT
sqlparse
tzdata